requirements.txt
flask
python-dotenv
google-genai
flask-sqlalchemy
flask-migrate
